update README, script names
