+++ README.md
@@ -1,6 +1,6 @@
 # SSV Network Project
 
-This repository contains a SSVRegistry and SSVNetwork smart contacts project.
+This repository contains a SSVNetwork smart contacts project.
 
 ## Quick start
 
@@ -23,7 +23,7 @@ npx hardhat node
 For more details about it and how to use MainNet forking you can find [here](https://hardhat.org/hardhat-network/).
 
 ### Compile contracts
-Take a look at `contracts/` folder, you should be able to find `SSVNetwork.sol`, `SSVNetworkV2.sol` as simple contract example.
+Take a look at `contracts/` folder, you should be able to find `SSVNetwork.sol`.
 To compile it, simply run:
 
 ```sh
@@ -34,7 +34,7 @@ npx hardhat compile
 
 ### Step 1: Test contracts
 Take a look at `test/` folder, you should be able to find tests related to specific actions.
-It comes with tests that use [Waffle](https://getwaffle.io/) and [Ethers.js](https://github.com/ethers-io/ethers.js/).
+It comes with tests that use [Ethers.js](https://github.com/ethers-io/ethers.js/).
 To run tests, run:
 
 ```sh
@@ -42,26 +42,30 @@ npx hardhat test
 ```
 
 ### Step 2: Deploy new contracts
-We use [Proxy Upgrade pattern](https://docs.openzeppelin.com/upgrades-plugins/1.x/proxies) for smart contracts to have an ability to upgrade them later.
-To deploy the contract we will use a Hardhat script. Inside `scripts/` you will find `ssv-registry-deploy.ts` and `ssv-network-deploy.ts` files.
-As general rule, you can target any network configured in the `hardhat.config.ts`
-
-#### Deploy SSV Registry
-
-```sh
-npx hardhat run --network <your-network> scripts/ssv-registry-deploy.ts
-```
-Output of this action will be smart SSV Registry contract proxy address.
+We use [UUPS Proxy Upgrade pattern](https://docs.openzeppelin.com/contracts/4.x/api/proxy) for smart contracts to have an ability to upgrade them later.
+To deploy the contract we will use a Hardhat script. Inside `scripts/` you will find `ssv-network-deploy.ts` and `ssv-network-upgrade.ts` files.
+As general rule, you can target any network configured in the `hardhat.config.ts`,
+specifying the right [network]_ETH_NODE_URL and [network]_OWNER_PRIVATE_KEY in `.env` file.
 
 #### Deploy SSV Network
+
 Before run the cli command, in `.env` need to add the following seetings:
 
 ```sh
-SSVREGISTRY_ADDRESS=#SSV Registry contract address
+[NETWORK]_ETH_NODE_URL=# RPC URL of the node
+[NETWORK]_OWNER_PRIVATE_KEY=# Private key of the deployer account, without 0x prefix
+GAS_PRICE=# example 30000000000
+GAS=# example 8000000
+ETHERSCAN_KEY=# etherescan API key
 SSVTOKEN_ADDRESS=#SSV Token contract address
+MINIMUM_BLOCKS_BEFORE_LIQUIDATION=# custom param
+OPERATOR_MAX_FEE_INCREASE=# custom param
+DECLARE_OPERATOR_FEE_PERIOD=# custom param
+EXECUTE_OPERATOR_FEE_PERIOD=# custom param
+VALIDATORS_PER_OPERATOR_LIMIT=# custom param
+REGISTERED_OPERATORS_PER_ACCOUNT_LIMIT=# custom param
+PROXY_ADDRESS=# Proxy address, set it when runnning ssv-network-upgrade.ts script
 ```
-
-After that:
 
 ```sh
 npx hardhat run --network <your-network> scripts/ssv-network-deploy.ts
@@ -70,8 +74,8 @@ Output of this action will be smart SSV Network contract proxy address.
 
 ### Step 3: Verify implementation contract on etherscan (each time after upgrade)
 Open `.openzeppelin/<network>.json` file and find `[impls.<hash>.address]` value which is implementation smart contract address.
-We use [Proxy Upgrade pattern](https://docs.openzeppelin.com/upgrades-plugins/1.x/proxies) for smart contracts to have an ability to upgrade them later.
-To deploy the contract we will use a Hardhat script. Inside `scripts/` you will find `deploy.ts` file.
+We use [UUPS Proxy Upgrade pattern](https://docs.openzeppelin.com/contracts/4.x/api/proxy) for smart contracts to have an ability to upgrade them later.
+To deploy the contract we will use a Hardhat script. Inside `scripts/` you will find `ssv-network-deploy.ts` file.
 Run this:
 ```sh
 npx hardhat  verify --network <network> <implementation-address>
@@ -87,18 +91,34 @@ The proxy\'s (<proxy-address) implementation contract is found at: <implementati
 To be sure that values are correct and click `Save` button. As result on etherscan proxy address page in `Contract` tab you will find two new buttons:
 `Write as Proxy` and `Read as Proxy` which will represent implementation smart contract functions interface and the actual state.
 
-### Step 5: Upgrade contract
+### Step 5: Upgrade SSVNetwork contract
 Once we have tested our new implementation, for example `contracts/SSVNetwork.sol` we can prepare the upgrade.
-This will validate and deploy our new implementation contract.
-#### Upgrade SSV Registry contract:
+
+In `.env` file, remember to set `PROXY_ADDRESS`.
+
+**Important**
+Pay special attention when changing storage layout, for example adding new storage variables
+in `SSVNetwork` (base) contract.
+There is a state variable `uint256[50] __gap;` that you should reduce the size according to
+the size of the new variables added. More info: [Storage Gaps](https://docs.openzeppelin.com/upgrades-plugins/1.x/writing-upgradeable#storage-gaps)
+
+To validate the upgrade before running it:
 ```sh
-PROXY_ADDRESS=0x... npx hardhat run --network <network> scripts/ssv-registry-upgrade.ts
+npx hardhat run --network <your-network> scripts/ssv-network-validate-upgrade.ts
+```
+To fire the upgrade process:
+```sh
+npx hardhat run --network <your-network> scripts/ssv-network-upgrade.ts
 ```
 
-#### Upgrade SSV Network contract:
-```sh
-PROXY_ADDRESS=0x... npx hardhat run --network <network> scripts/ssv-network-upgrade.ts
-```
+If you get the error:
+`
+Error: invalid hex string ...
+reason: 'invalid hex string',
+code: 'INVALID_ARGUMENT',
+`
+Set or change the parameters `GAS_PRICE` and `GAS` in `.env` file.
+
 ### dApp UI to interact with smart contract
 
 ```sh
@@ -106,29 +126,3 @@ https://eth95.dev/?network=1&address=0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0
 ```
 
 UI dApp [direct link](https://eth95.dev/?network=1&address=0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0)
-
-## User Guide
-
-You can find detailed instructions on using this repository and many tips in [its documentation](https://hardhat.org/tutorial).
-For a complete introduction to Hardhat, refer to [this guide](https://hardhat.org/getting-started/#overview).
-
-## What’s Included?
-
-Your environment will have everything you need to build a Dapp powered by Hardhat and React.
-
-- [Hardhat](https://hardhat.org/): An Ethereum development task runner and testing network.
-- [Mocha](https://mochajs.org/): A JavaScript test runner.
-- [Chai](https://www.chaijs.com/): A JavaScript assertion library.
-- [ethers.js](https://docs.ethers.io/ethers.js/html/): A JavaScript library for interacting with Ethereum.
-- [Waffle](https://github.com/EthWorks/Waffle/): To have Ethereum-specific Chai assertions/mathers.
-
-## Security Audit [OLD Version]
-
-Full audit report [CoinFabrik Report](./docs/SSV_Token_Dex&Vesting_audit.pdf)
-
-## Troubleshooting
-
-- `Invalid nonce` errors: if you are seeing this error on the `npx hardhat node`
-  console, try resetting your Metamask account. This will reset the account's
-  transaction history and also the nonce. Open Metamask, click on your account
-  followed by `Settings > Advanced > Reset Account`.
