+++ hardhat.config.ts
@@ -19,7 +19,7 @@ const config: HardhatUserConfig = {
         settings: {
           optimizer: {
             enabled: true,
-            runs: 500
+            runs: 800
           }
         }
       }
